% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installers.R
\name{binary_download}
\alias{binary_download}
\title{Download Binary Packages}
\usage{
binary_download(
  url,
  binary_file_name = base::basename(url),
  verbose = TRUE,
  mode = "wb",
  timeout = 600
)
}
\arguments{
\item{url}{A link containing the binary file to download.}

\item{binary_file_name}{Name of the binary file to save. Default \code{basname(url)}}

\item{verbose}{Display a status messages. Default \code{true}}

\item{mode}{Mode to use for download. Default is "wb" (binary mode)}

\item{timeout}{Timeout in seconds. Default is 600 (10 minutes)}
}
\value{
The file path for the binary file in the temporary \emph{R} directory
}
\description{
The \emph{R} temporary directory is created using \code{\link[=tempdir]{tempdir()}} that
consults the environment values of \link{TMPDIR}, \href{TMPDIR}{TMP}, or \href{TMPDIR}{TEMP} to establish
the path. We need to use a temporary directory as \code{root} is not guaranteed
to have access to user's files.
}
\details{
Downloads a binary package onto a user's computer into the temporary directory.
}
