% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{caller_env}
\alias{caller_env}
\title{Get Caller Environment}
\usage{
caller_env(n = 1)
}
\arguments{
\item{n}{Number of frames to go back}
}
\value{
Caller environment
}
\description{
Get Caller Environment
}
\keyword{internal}
