% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-custom.R
\name{cli_danger}
\alias{cli_danger}
\title{Display a danger/important message}
\usage{
cli_danger(..., .envir = parent.frame())
}
\arguments{
\item{...}{Message parts, passed to cli::format_inline}

\item{.envir}{Environment to evaluate in}
}
\description{
Display a danger/important message
}
\keyword{internal}
