% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-custom.R
\name{cli_error}
\alias{cli_error}
\title{Throw an error with a formatted message}
\usage{
cli_error(..., .envir = parent.frame(), call = caller_env(), advice = NULL)
}
\arguments{
\item{...}{Message parts, passed to cli::format_inline}

\item{.envir}{Environment to evaluate in}

\item{call}{The calling environment}

\item{advice}{Optional advice to provide to the user}
}
\description{
Throw an error with a formatted message
}
\keyword{internal}
