% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-custom.R
\name{cli_info}
\alias{cli_info}
\title{Display an info message}
\usage{
cli_info(..., .envir = parent.frame())
}
\arguments{
\item{...}{Message parts, passed to cli::format_inline}

\item{.envir}{Environment to evaluate in}
}
\description{
Display an info message
}
\keyword{internal}
