% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-custom.R
\name{cli_process_done}
\alias{cli_process_done}
\title{Complete a process}
\usage{
cli_process_done(id, msg = NULL)
}
\arguments{
\item{id}{The progress bar ID}

\item{msg}{Optional completion message}
}
\description{
Complete a process
}
\keyword{internal}
