% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-custom.R
\name{cli_process_start}
\alias{cli_process_start}
\title{Display the start of a process}
\usage{
cli_process_start(msg, .envir = parent.frame(), total = 100, clear = FALSE)
}
\arguments{
\item{msg}{The process description}

\item{.envir}{Environment to evaluate in}

\item{total}{The total number of steps (default: 100)}

\item{clear}{Whether to clear the progress bar when done (default: FALSE)}
}
\value{
The ID of the progress bar
}
\description{
Display the start of a process
}
\keyword{internal}
