% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-custom.R
\name{cli_process_update}
\alias{cli_process_update}
\title{Update a process progress}
\usage{
cli_process_update(id, ratio, status_msg = NULL)
}
\arguments{
\item{id}{The progress bar ID}

\item{ratio}{The progress ratio (0-1) or number of steps completed}

\item{status_msg}{Optional status message to display}
}
\description{
Update a process progress
}
\keyword{internal}
