% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-custom.R
\name{cli_success}
\alias{cli_success}
\title{Display a success message}
\usage{
cli_success(..., .envir = parent.frame())
}
\arguments{
\item{...}{Message parts, passed to cli::format_inline}

\item{.envir}{Environment to evaluate in}
}
\description{
Display a success message
}
\keyword{internal}
