% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-custom.R
\name{cli_system_update}
\alias{cli_system_update}
\title{Display a detailed system update}
\usage{
cli_system_update(title, items)
}
\arguments{
\item{title}{Title of the update section}

\item{items}{Vector of items to display as a bulleted list}
}
\description{
Display a detailed system update
}
\keyword{internal}
