% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-custom.R
\name{cli_waiting}
\alias{cli_waiting}
\title{Display a waiting message}
\usage{
cli_waiting(..., .envir = parent.frame())
}
\arguments{
\item{...}{Message parts, passed to cli::format_inline}

\item{.envir}{Environment to evaluate in}
}
\description{
Display a waiting message
}
\keyword{internal}
