% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-custom.R
\name{cli_warning}
\alias{cli_warning}
\title{Display a warning message}
\usage{
cli_warning(..., .envir = parent.frame())
}
\arguments{
\item{...}{Message parts, passed to cli::format_inline}

\item{.envir}{Environment to evaluate in}
}
\description{
Display a warning message
}
\keyword{internal}
