% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-custom.R
\name{format_cmd}
\alias{format_cmd}
\title{Format a command for display}
\usage{
format_cmd(cmd)
}
\arguments{
\item{cmd}{The command to format}
}
\value{
A formatted string with code styling
}
\description{
Format a command for display
}
\keyword{internal}
