% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-custom.R
\name{format_msg}
\alias{format_msg}
\title{Format a message with a package prefix}
\usage{
format_msg(msg, .envir = parent.frame())
}
\arguments{
\item{msg}{The message to format}

\item{.envir}{Environment to evaluate in}
}
\value{
A formatted string with package prefix
}
\description{
Format a message with a package prefix
}
\keyword{internal}
