% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{is_aarch64}
\alias{is_aarch64}
\title{Check if System is aarch64}
\usage{
is_aarch64()
}
\value{
TRUE if system is Apple Silicon (M-series) Mac, FALSE otherwise
}
\description{
Check if System is aarch64
}
\keyword{internal}
