% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{is_macos_big_sur}
\alias{is_macos_big_sur}
\title{Check if macOS Big Sur}
\usage{
is_macos_big_sur()
}
\value{
TRUE if system is macOS Big Sur, FALSE otherwise
}
\description{
Check if macOS Big Sur
}
\keyword{internal}
