% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{is_macos_catalina}
\alias{is_macos_catalina}
\title{Check if macOS Catalina}
\usage{
is_macos_catalina()
}
\value{
TRUE if system is macOS Catalina, FALSE otherwise
}
\description{
Check if macOS Catalina
}
\keyword{internal}
