% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{is_macos_high_sierra}
\alias{is_macos_high_sierra}
\title{Check if macOS High Sierra}
\usage{
is_macos_high_sierra()
}
\value{
TRUE if system is macOS High Sierra, FALSE otherwise
}
\description{
Check if macOS High Sierra
}
\keyword{internal}
