% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{is_macos_mojave}
\alias{is_macos_mojave}
\title{Check if macOS Mojave}
\usage{
is_macos_mojave()
}
\value{
TRUE if system is macOS Mojave, FALSE otherwise
}
\description{
Check if macOS Mojave
}
\keyword{internal}
