% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{is_macos_monterey}
\alias{is_macos_monterey}
\title{Check if macOS Monterey}
\usage{
is_macos_monterey()
}
\value{
TRUE if system is macOS Monterey, FALSE otherwise
}
\description{
Check if macOS Monterey
}
\keyword{internal}
