% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{is_macos_r_supported}
\alias{is_macos_r_supported}
\title{Check if macOS Version is Supported for R}
\usage{
is_macos_r_supported()
}
\value{
TRUE if macOS version is supported, FALSE otherwise
}
\description{
Check if macOS Version is Supported for R
}
\keyword{internal}
