% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{is_macos_sequoia}
\alias{is_macos_sequoia}
\title{Check if macOS Sequoia}
\usage{
is_macos_sequoia()
}
\value{
TRUE if system is macOS Sequoia, FALSE otherwise
}
\description{
Check if macOS Sequoia
}
\keyword{internal}
