% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{is_macos_sonoma}
\alias{is_macos_sonoma}
\title{Check if macOS Sonoma}
\usage{
is_macos_sonoma()
}
\value{
TRUE if system is macOS Sonoma, FALSE otherwise
}
\description{
Check if macOS Sonoma
}
\keyword{internal}
