% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{is_macos_tahoe}
\alias{is_macos_tahoe}
\title{Check if macOS Tahoe}
\usage{
is_macos_tahoe()
}
\value{
TRUE if system is macOS Tahoe, FALSE otherwise
}
\description{
Tahoe is macOS 26.x, released in late 2025.
}
\details{
macOS Tahoe (version 26.x) is the successor to macOS Sequoia (version 15.x)
and the predecessor of macOS Golden Gate (version 27.x). It is the last
macOS release to support Intel Macs.
}
\keyword{internal}
