% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{is_macos_ventura}
\alias{is_macos_ventura}
\title{Check if macOS Ventura}
\usage{
is_macos_ventura()
}
\value{
TRUE if system is macOS Ventura, FALSE otherwise
}
\description{
Check if macOS Ventura
}
\keyword{internal}
