% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{is_r_version_at_least}
\alias{is_r_version_at_least}
\title{Check if the Running R Version is At Least a Target}
\usage{
is_r_version_at_least(target, version = r_version_major_minor())
}
\arguments{
\item{target}{Target R version (e.g. \code{"4.3"}) to compare against.}

\item{version}{The version to test, defaults to the running R version.}
}
\value{
TRUE if \code{version} is greater than or equal to \code{target}, FALSE otherwise
}
\description{
Check if the Running R Version is At Least a Target
}
\keyword{internal}
