% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{is_r_version_supported}
\alias{is_r_version_supported}
\title{Check if the Running R Version is Supported}
\usage{
is_r_version_supported(version = r_version_major_minor())
}
\arguments{
\item{version}{The version to test, defaults to the running R version.}
}
\value{
TRUE if \code{version} falls within the supported window (inclusive),
FALSE otherwise
}
\description{
Check if the Running R Version is Supported
}
\keyword{internal}
