% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{is_x86_64}
\alias{is_x86_64}
\title{Check if System is x86_64}
\usage{
is_x86_64()
}
\value{
TRUE if system is Intel-based Mac, FALSE otherwise
}
\description{
Check if System is x86_64
}
\keyword{internal}
