% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolchain.R
\name{macos_rtools_install}
\alias{macos_rtools_install}
\alias{macos_rtools_uninstall}
\title{Install and Uninstall the macOS R Toolchain}
\usage{
macos_rtools_install(
  password = base::getOption("macrtools.password"),
  verbose = TRUE
)

macos_rtools_uninstall(
  password = base::getOption("macrtools.password"),
  verbose = TRUE
)
}
\arguments{
\item{password}{Password for user account to request \code{sudo} access.}

\item{verbose}{Describe each step taken. Default \code{TRUE}}
}
\description{
The \code{macos_rtools_install()} function aims to install all required dependencies
for the macOS R compilation toolchain. Meanwhile, the \code{macros_rtools_uninstall()}
function aims to remove any installed files from your computer.
}
\details{
The macOS R compilation toolchain consists of:
\enumerate{
\item Xcode CLI
\item gfortran
\item A series of binary packages from the \href{https://github.com/R-macos/recipes}{\code{recipes}} system to compile R.
}

The \code{mac_rtools_install()} function attempts to install each of the required
components. If we detect that the Xcode.app IDE is installed, we'll skip
attempting to install the Xcode CLI software.

Meanwhile, the \code{mac_rtools_uninstall()} function aims to
delete or uninstall the Xcode CLI and gfortran binaries. At the present moment,
there is no support for uninstalling the binary packages from \code{recipes}.
}
\section{Headless or Unattended Installation}{


For an installation that does not require a human in the loop, please set
the environment variable \code{macrtools.password} for the user profile:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{Sys.setenv("macrtools.password" = "your-password-here")
}\if{html}{\out{</div>}}
}

\examples{

\dontrun{

# Install all tools required for compiling code on macOS with R
macos_rtools_install()

# Enter a password to avoid being prompted!
macos_rtools_install(password = "hello-compiled-code-world")

}

\dontrun{

# Remove Xcode CLI and gfortran tools required for compiling code on macOS with R
macos_rtools_uninstall()

# Enter a password to avoid being prompted!
macos_rtools_uninstall(password = "hello-compiled-code-world")

}
}
