% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{newest_supported_macos_name}
\alias{newest_supported_macos_name}
\title{Marketing Name of the Newest Supported macOS Release}
\usage{
newest_supported_macos_name()
}
\value{
The marketing name of the newest supported macOS release.
}
\description{
Marketing Name of the Newest Supported macOS Release
}
\keyword{internal}
