% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null_coalesce}
\alias{null_coalesce}
\alias{\%||\%}
\title{Null Coalesce Operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First value (may be NULL)}

\item{y}{Default value if x is NULL}
}
\value{
x if not NULL, otherwise y
}
\description{
Null Coalesce Operator
}
\keyword{internal}
