% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell.R
\name{shell_command}
\alias{shell_command}
\title{Execute shell commands}
\usage{
shell_command(cmd, verbose = TRUE)
}
\arguments{
\item{cmd}{The command to execute}

\item{verbose}{Display the command being executed}
}
\value{
The exit status of the command (0 for success)
}
\description{
Functions to execute shell commands with or without sudo
}
\keyword{internal}
