% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell.R
\name{shell_sudo_command}
\alias{shell_sudo_command}
\title{Execute shell command with sudo}
\usage{
shell_sudo_command(cmd, password, verbose = TRUE, prefix = "sudo -kS ")
}
\arguments{
\item{cmd}{The command to execute}

\item{password}{User password for sudo privileges}

\item{verbose}{Display the command being executed}

\item{prefix}{The sudo prefix (default: "sudo -kS ")}
}
\value{
The exit status of the command (0 for success)
}
\description{
Execute shell command with sudo
}
\keyword{internal}
