% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installers.R
\name{tar_package_download}
\alias{tar_package_download}
\alias{tar_package_install}
\title{Download and Install Tar Packages}
\usage{
tar_package_download(url, binary_file_name = basename(url), verbose = TRUE)

tar_package_install(
  path_to_tar,
  install_directory,
  strip_levels,
  sudo = TRUE,
  password = NULL,
  verbose = TRUE
)
}
\arguments{
\item{url}{A link containing the binary tar file to download.}

\item{binary_file_name}{Name of the binary file to save}

\item{verbose}{Display status messages}

\item{path_to_tar}{Location of where the tar file is}

\item{install_directory}{Location of where to unpack or extract the tar file.}

\item{strip_levels}{Remove nesting inside of the \code{tar} file}

\item{sudo}{Run the command as \code{root} through \code{sudo}. Default \code{TRUE}}

\item{password}{User password to use to enter \code{sudo} mode.}
}
\value{
The downloaded tar file path
}
\description{
The \emph{R} temporary directory is created using \code{\link[=tempdir]{tempdir()}} that
consults the environment values of \link{TMPDIR}, \link{TMP}, or \link{TEMP} to establish
the path. We need to use a temporary directory as \code{root} is not guaranteed
to have access to user's files.

We need to use \code{sudo} access to unpack the binary into the low-level system
holding area.
}
\details{
Downloads or installs a tar package onto a user's computer.
}
