% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installers.R
\name{tar_package_install}
\alias{tar_package_install}
\title{Install Binary Package in a Tar Format}
\usage{
tar_package_install(
  path_to_tar,
  install_directory,
  strip_levels,
  sudo = TRUE,
  password = NULL,
  verbose = TRUE
)
}
\arguments{
\item{path_to_tar}{Location of where the tar file is}

\item{install_directory}{Location of where to unpack or extract the tar file.}

\item{strip_levels}{Remove nesting inside of the \code{tar} file}

\item{sudo}{Run the command as \code{root} through \code{sudo}. Default \code{TRUE}}

\item{password}{User password to use to enter \code{sudo} mode.}

\item{verbose}{Display status messages}
}
\description{
The use of \code{sudo} is required to unpack the binary file
into the low-level system area.
}
\details{
Unpacks the Tar package and places it into a system library
}
