% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{timestamp_now}
\alias{timestamp_now}
\title{Current Timestamp as a Formatted String}
\usage{
timestamp_now()
}
\value{
The current time formatted as \code{"\%Y-\%m-\%d \%H:\%M:\%S"}.
}
\description{
Current Timestamp as a Formatted String
}
\keyword{internal}
