% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{version_above}
\alias{version_above}
\title{Check if Version is Above Threshold}
\usage{
version_above(software_version, than)
}
\arguments{
\item{software_version}{Version string to check}

\item{than}{Threshold version to compare against}
}
\value{
TRUE if software_version is above than, FALSE otherwise
}
\description{
Check if Version is Above Threshold
}
\keyword{internal}
