% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{version_between}
\alias{version_between}
\title{Check if Version is Between Bounds}
\usage{
version_between(software_version, lower, greater_strict)
}
\arguments{
\item{software_version}{Version string to check}

\item{lower}{Lower bound for version check (inclusive)}

\item{greater_strict}{Upper bound for version check (exclusive)}
}
\value{
TRUE if software_version is between bounds, FALSE otherwise
}
\description{
Check if Version is Between Bounds
}
\keyword{internal}
