% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfortran.R, R/xcode-cli.R
\name{gfortran_version}
\alias{gfortran_version}
\alias{xcode_select}
\alias{xcode_select_path}
\alias{xcode_select_version}
\alias{print.xcodeselect}
\title{Interface with \code{xcode-select} Shell Commands}
\usage{
gfortran_version()

xcode_select(args)

xcode_select_path()

xcode_select_version()

\method{print}{xcodeselect}(x, ...)
}
\arguments{
\item{args}{Flag arguments to pass to \code{xcode-select}}

\item{x}{An object with class \code{xcodeselect}}

\item{...}{Additional parameters}
}
\description{
Trigger \code{xcode-select} commands from within \emph{R}
}
