% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcode.R
\name{xcode_app_version}
\alias{xcode_app_version}
\title{Version of an Xcode Bundle}
\usage{
xcode_app_version(bundle)
}
\arguments{
\item{bundle}{Path to an \code{Xcode.app} bundle.}
}
\value{
The bundle's short version string, or \code{NA_character_}.
}
\description{
Version of an Xcode Bundle
}
\keyword{internal}
