% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcodebuild.R
\name{xcodebuild}
\alias{xcodebuild}
\alias{xcodebuild_version}
\title{Interface with \code{xcodebuild} Shell Commands}
\usage{
xcodebuild(args)

xcodebuild_version()
}
\arguments{
\item{args}{Flag arguments to pass to \code{xcodebuild}}
}
\description{
Trigger \code{xcodebuild} commands from within \emph{R}
}
