% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfortran.R
\name{is_gfortran_installed}
\alias{is_gfortran_installed}
\alias{gfortran_version}
\alias{gfortran_install}
\alias{gfortran_uninstall}
\alias{gfortran_update}
\title{Find, Install, or Uninstall gfortran}
\usage{
is_gfortran_installed()

gfortran_version()

gfortran_install(password = getOption("macrtools.password"), verbose = TRUE)

gfortran_uninstall(password = getOption("macrtools.password"), verbose = TRUE)

gfortran_update(password = getOption("macrtools.password"), verbose = TRUE)
}
\arguments{
\item{password}{Password for user account to install software. Default is
\code{NULL}.}

\item{verbose}{Display messages indicating status}
}
\description{
Set of functions that seek to identify whether gfortran was installed,
allow gfortran to be installed, and removing gfortran.
}
\section{Check if \code{gfortran} is installed}{


Checks using the \code{gfortran --version} command to see if a response is present.
}

\section{Install \code{gfortran}}{

The \code{gfortran_install()} function aims to install \code{gfortran} into
two locations:
\itemize{
\item Intel (\code{x86_64})
\itemize{
\item \verb{/usr/local/gfortran}
\item \verb{/usr/local/bin/gfortran}
}
\item M1/M2 (\code{aarch64})
\itemize{
\item \verb{/opt/r/arm64/gfortran/}
\item \verb{/opt/r/arm64/bin/gfortran}
}
}

The Intel \code{gfortran} installer is a DMG image that is mounted, installed,
and unmounted.

The M1/M2 \code{gfortran} installer is a tar file that is unpacked into the
directory.
}

\section{Uninstalling \code{gfortran}}{

The \code{gfortran_uninstall()} attempts to remove \code{gfortran} from
two locations:
\itemize{
\item Intel (\code{x86_64})
\itemize{
\item \verb{/usr/local/gfortran}
\item \verb{/usr/local/bin/gfortran}
}
\item M1/M2 (\code{aarch64})
\itemize{
\item \verb{/opt/r/arm64/gfortran/}
\item \verb{/opt/r/arm64/bin/gfortran}
}
}

Using the \emph{R} sanitized \emph{shell} command of:

\if{html}{\out{<div class="sourceCode sh">}}\preformatted{sudo -kS rm -r /usr/local/gfortran /usr/local/bin/gfortran
}\if{html}{\out{</div>}}

Or:

\if{html}{\out{<div class="sourceCode sh">}}\preformatted{sudo -kS rm -r /opt/r/arm64/gfortran/ /opt/r/arm64/bin/gfortran
}\if{html}{\out{</div>}}

These uninstall steps are based on:

\url{https://gcc.gnu.org/wiki/GFortranBinariesMacOS}

And the desired path for \code{aarch64}:

\url{https://mac.r-project.org/tools/}
}

\section{Update \code{gfortran}}{

The \code{gfortran_update()} attempts to update the version of \code{gfortran} installed
using the provided \code{gfortran-update-sdk} inside of \verb{/opt/R/arm64/gfortran/bin}.

Note: This update command only works for M1/M2 (\code{arm64}) users on \emph{R} 4.2
or above.
}

\examples{
# Check if gfortran is installed
is_gfortran_installed()
}
