% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfortran.R
\name{is_gfortran_installed}
\alias{is_gfortran_installed}
\alias{gfortran_install}
\alias{gfortran_uninstall}
\alias{gfortran_update}
\title{Find, Install, or Uninstall gfortran}
\usage{
is_gfortran_installed()

gfortran_install(verbose = TRUE)

gfortran_uninstall(verbose = TRUE)

gfortran_update(verbose = TRUE)
}
\description{
Set of functions that seek to identify whether gfortran was installed,
allow gfortran to be installed, and removing gfortran.
}
\section{Check if \code{gfortran} is installed}{


Checks using the \code{gfortran --version} command to see if a response is present.
}

\examples{
# Check if gfortran is installed
is_gfortran_installed()
}
