% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{install_directory_x86_64}
\alias{install_directory_x86_64}
\alias{install_directory_x86_64_r_version_4_3}
\alias{install_directory_arm64}
\alias{install_directory_xcode_cli}
\alias{install_directory_xcode_app}
\title{Installation Directory Locations}
\usage{
install_directory_x86_64()

install_directory_x86_64_r_version_4_3()

install_directory_arm64()

install_directory_xcode_cli()

install_directory_xcode_app()
}
\value{
A string containing a fixed path to the installation directory.
}
\description{
Pre-defined constants for each software install location.
}
\details{
We have the following fixed paths:
\itemize{
\item Intel (x86_64)
\itemize{
\item macOS 10.13: \verb{/usr/local}
\item macOS 11: \verb{/opt/R/x86_64}
}
\item M1, M2, M3 (\code{arm64} or \code{aarch64})
\itemize{
\item \verb{/opt/R/arm64}
}
\item Xcode Command Line Tools (Xcode CLI)
\itemize{
\item \verb{/Library/Developer/CommandLineTools}
}
\item Xcode App
\itemize{
\item \verb{/Applications/Xcode.app/Contents/Developer}
}
}
}
