% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{is_macos_sequoia}
\alias{is_macos_sequoia}
\title{Check if macOS Sequoia}
\usage{
is_macos_sequoia()
}
\value{
TRUE if system is macOS Sequoia, FALSE otherwise
}
\description{
Sequoia is macOS 15.x, released in late 2024.
}
\details{
macOS Sequoia (version 15.x) is the successor to macOS
Sonoma (version 14.x).
}
\keyword{internal}
