% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{is_macos_sonoma}
\alias{is_macos_sonoma}
\title{Check if macOS Sonoma}
\usage{
is_macos_sonoma()
}
\value{
TRUE if system is macOS Sonoma, FALSE otherwise
}
\description{
Sonoma is macOS 14.x, released in late 2023.
}
\details{
macOS Sonoma (version 14.x) is the successor to macOS
Ventura (version 13.x).
}
\keyword{internal}
