% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{macos_version_in_range}
\alias{macos_version_in_range}
\title{Check if the macOS Version Falls in a Range}
\usage{
macos_version_in_range(lower, upper)
}
\arguments{
\item{lower}{Lower bound for the macOS version (inclusive).}

\item{upper}{Upper bound for the macOS version (exclusive).}
}
\value{
TRUE if the running macOS version is in \verb{[lower, upper)}.
}
\description{
Check if the macOS Version Falls in a Range
}
\keyword{internal}
