% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmp.R
\name{is_openmp_installed}
\alias{is_openmp_installed}
\alias{openmp_version}
\alias{openmp_install}
\alias{openmp_uninstall}
\alias{openmp_test}
\title{Find, Install, or Uninstall OpenMP}
\usage{
is_openmp_installed()

openmp_version()

openmp_install(
  password = base::getOption("macrtools.password"),
  verbose = TRUE,
  configure_makevars = TRUE
)

openmp_uninstall(
  password = base::getOption("macrtools.password"),
  verbose = TRUE,
  configure_makevars = TRUE
)

openmp_test()
}
\arguments{
\item{password}{Password for user account to install software. Default is
\code{NULL}.}

\item{verbose}{Display messages indicating status}

\item{configure_makevars}{Automatically configure ~/.R/Makevars with or without OpenMP flags.
Default is \code{TRUE}.}
}
\description{
Set of functions that seek to identify whether OpenMP was installed,
allow OpenMP to be installed, and removing OpenMP.
}
\details{
The \verb{openmp_*()} suite of functions attempts to locate, install, and uninstall
OpenMP runtime library based on the Xcode version. OpenMP is installed to:
\itemize{
\item Library: \verb{/usr/local/lib/libomp.dylib}
\item Headers: \verb{/usr/local/include/omp.h}, \verb{/usr/local/include/ompt.h}, \verb{/usr/local/include/omp-tools.h}
}

Apple has explicitly disabled OpenMP support in Xcode compilers, but the
runtime library can be installed separately and used with \code{-Xclang -fopenmp}
compiler flags.
}
\section{Check if \code{OpenMP} is installed}{


Checks the local file system for whether OpenMP runtime library is installed
in the default installation location.
}

\section{Installing \code{OpenMP}}{

The \code{openmp_install()} function aims to install the appropriate OpenMP
runtime library based on the detected Xcode version.
\subsection{OpenMP Installation Process}{

The installation process automatically detects your Xcode version and
downloads the corresponding OpenMP runtime from the R-project repository:

\if{html}{\out{<div class="sourceCode sh">}}\preformatted{# Download the appropriate version
curl -O https://mac.r-project.org/openmp/openmp-VERSION-darwin20-Release.tar.gz

# Install to system directories
sudo tar fvxz openmp-VERSION-darwin20-Release.tar.gz -C /
}\if{html}{\out{</div>}}

The installation includes:
\itemize{
\item \verb{/usr/local/lib/libomp.dylib} - Main OpenMP runtime library
\item \verb{/usr/local/include/omp.h} - OpenMP API header
\item \verb{/usr/local/include/ompt.h} - OpenMP Tools interface header
\item \verb{/usr/local/include/omp-tools.h} - OpenMP Tools header
}
}

\subsection{Using OpenMP in R Packages}{

After installation, you can enable OpenMP support in R packages by adding
the following to your \verb{~/.R/Makevars} file:

\if{html}{\out{<div class="sourceCode sh">}}\preformatted{CPPFLAGS += -Xclang -fopenmp
LDFLAGS += -lomp
}\if{html}{\out{</div>}}

Or install packages with:

\if{html}{\out{<div class="sourceCode sh">}}\preformatted{PKG_CPPFLAGS='-Xclang -fopenmp' PKG_LIBS=-lomp R CMD INSTALL myPackage
}\if{html}{\out{</div>}}
}
}

\section{Uninstalling \code{OpenMP}}{


The \code{openmp_uninstall()} attempts to remove OpenMP from the default
installation locations.

We use the \emph{R} sanitized \emph{shell} command to remove the installed files:

\if{html}{\out{<div class="sourceCode sh">}}\preformatted{sudo rm -f /usr/local/lib/libomp.dylib
sudo rm -f /usr/local/include/omp.h
sudo rm -f /usr/local/include/ompt.h
sudo rm -f /usr/local/include/omp-tools.h
sudo rm -f /usr/local/include/ompx.h
}\if{html}{\out{</div>}}

Note that \code{ompx.h} was included in LLVM 19.1.0 and
may not exist in older OpenMP versions.
}

\section{Testing OpenMP Installation}{

After installing OpenMP, you can test if it's working by creating a simple
test package or by adding the following to your \verb{~/.R/Makevars} file:

\if{html}{\out{<div class="sourceCode sh">}}\preformatted{CPPFLAGS += -Xclang -fopenmp
LDFLAGS += -lomp
}\if{html}{\out{</div>}}

Then try installing a package that uses OpenMP, such as:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{install.packages("data.table", type = "source")
}\if{html}{\out{</div>}}
}

\examples{
# Check if OpenMP is installed
is_openmp_installed()
}
