% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.cli}
\alias{print.cli}
\title{Print CLI Responses}
\usage{
\method{print}{cli}(x, ...)
}
\arguments{
\item{x}{An object with class \code{cli}}

\item{...}{Additional parameters}
}
\description{
Print CLI Responses
}
