% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes.R
\name{recipes_binary_install}
\alias{recipes_binary_install}
\title{Install Binary Library from the CRAN R macOS Recipes Project}
\usage{
recipes_binary_install(
  pkgs,
  url = "https://mac.R-project.org/bin",
  os = tolower(paste0(system("uname -s", intern = TRUE), gsub("\\\\..*", "",
    system("uname -r", intern = TRUE)))),
  arch = system("uname -m", intern = TRUE),
  os.arch = "auto",
  dependencies = TRUE,
  action = c("install", "list"),
  sudo = TRUE,
  password = NULL,
  verbose = TRUE
)
}
\arguments{
\item{pkgs}{Character vector of binary names to install, \code{"all"} for all binaries, or \code{"r-base-dev"} for \emph{R} binaries.}

\item{url}{URL of the repository root. Default \url{https://mac.R-project.org/bin}}

\item{os}{Name and version of the OS, e.g. \code{"darwin22"} where \code{"darwin"}
refers to macOS and \code{22} is the kernel version number.}

\item{arch}{The architecture of either \code{arm64} (M1/M2) or \code{x86_64} (Intel). This is only used if \code{os.arch="auto"}.}

\item{os.arch}{Either name of the repository such as \code{"darwin20/arm64"}, \code{"darwin17/x86_64"}, or \code{"auto"}.
Default \code{"auto"}.}

\item{dependencies}{Install build dependencies (\code{TRUE}) or only the requested packages (\code{FALSE}). Default \code{TRUE}.}

\item{action}{Determine if the binary should be downloaded and installed (\code{"install"})
or displayed (\code{"list"}). Default \code{"install"} to download and install the binaries.}

\item{sudo}{Attempt to install the binaries using \code{sudo} permissions.
Default \code{TRUE}.}

\item{password}{User password to switch into the \code{sudo} user. Default \code{NULL}.}

\item{verbose}{Describe the steps being taken. Default \code{TRUE}}
}
\description{
Convenience function that seeks to install pre-built binary libraries
used on \href{https://cran.r-project.org/}{CRAN} for macOS through the
\href{https://github.com/s-u/recipes}{recipes} system designed by Simon Urbanek.
}
\details{
The function attempts to detect the appropriate repository and installation
path for the binary packages when \code{"auto"} is set. By default, the
repository and the install path are either:\tabular{lll}{
   Name \tab Installation Location \tab Target \cr
   \href{https://mac.r-project.org/bin/darwin17/x86_64}{darwin17/x86_64} \tab /usr/local \tab macOS 10.13, Intel (x86_64) \cr
   \href{https://mac.r-project.org/bin/darwin20/x86_64}{darwin20/x86_64} \tab /opt/R/x86_64 \tab macOS 11, Intel (x86_64) \cr
   \href{https://mac.r-project.org/bin/darwin20/arm64}{darwin20/arm64} \tab /opt/R/arm64 \tab macOS 11, Apple M1 (arm64) \cr
}
}
\examples{
# Perform a dry-run to see the required development packages.
recipes_binary_install("r-base-dev", action = "list")

\dontrun{
# Install the mandatory library binaries for building R on macOS using sudo
recipes_binary_install("r-base-dev", sudo = TRUE)
}
}
\author{
Simon Urbanek wrote the function and made it available at
\url{https://mac.r-project.org/bin/}

James Joseph Balamuta packaged the function and added the option to use
\code{sudo} on the command line.
}
