% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcode.R
\name{is_xcode_app_installed}
\alias{is_xcode_app_installed}
\title{Detect if the Xcode.app IDE is Installed}
\usage{
is_xcode_app_installed()
}
\description{
Checks to see whether Xcode.app Integrated Developer Environment
(IDE) was installed and is active under the default location.
}
\details{
Checks to see if Xcode.app IDE is active by running:

\if{html}{\out{<div class="sourceCode sh">}}\preformatted{xcode-select -p
}\if{html}{\out{</div>}}

If it returns:

\if{html}{\out{<div class="sourceCode sh">}}\preformatted{/Applications/Xcode.app/Contents/Developer
}\if{html}{\out{</div>}}

We consider the full Xcode.app to be \strong{installed} and \strong{selected} as
the command line tools to use.
}
\examples{
# Check if Xcode.app IDE is on the path
is_xcode_app_installed()
}
