% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcode-cli.R
\name{is_xcode_cli_installed}
\alias{is_xcode_cli_installed}
\alias{xcode_cli_install}
\alias{xcode_cli_uninstall}
\title{Find, Install, or Uninstall XCode CLI}
\usage{
is_xcode_cli_installed()

xcode_cli_install(verbose = TRUE)

xcode_cli_uninstall(verbose = TRUE)
}
\arguments{
\item{verbose}{Display status messages}
}
\description{
Set of functions that seek to identify whether XCode CLI was installed,
allow XCode CLI to be installed, and removing XCode CLI.
}
\section{XCode CLI Installation Check}{


Checks using the \code{xcode-select -p} command to obtain the relevant
the directory. If the status code returned is 0, then the tools exist.
Otherwise, the return value will be 2 indicating that the tools are missing.

Implementation is based on:

\url{https://stackoverflow.com/questions/15371925/how-to-check-if-xcode-command-line-tools-are-installed}
}

