% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcode-cli.R
\name{xcode_select}
\alias{xcode_select}
\alias{xcode_select_path}
\alias{xcode_select_version}
\title{Interface with \code{xcode-select} Shell Commands}
\usage{
xcode_select(args)

xcode_select_path()

xcode_select_version()
}
\arguments{
\item{args}{Flag arguments to pass to \code{xcode-select}}
}
\description{
Trigger \code{xcode-select} commands from within \emph{R}
}
